Add tracing to notifications service.
This enables tracing to the notifications service.
